requirements.txt
python-telegram-bot==20.7
asyncio-mqtt==0.16.2
aiohttp==3.9.1
python-dotenv==1.0.0
asyncpg==0.29.0
redis==5.0.1
qrcode==7.4.2
Pillow==10.1.0
cryptography==41.0.7
pytz==2023.3
